first occourence CRUD

+++ src/services/projectServices.ts
@@ -1,55 +1,53 @@
+// ===== GERAL IMPORTS =====
 import { toast } from "react-toastify";
 import { db } from '../firebaseConfig/config'
-import { addDoc, deleteDoc, collection, doc, getDoc, getDocs, updateDoc, onSnapshot, setDoc, query, where } from 'firebase/firestore'
-import { getCurrentUser } from "./authService";
-import { deletePrototype, type PrototypeProps } from "./prototypeServices";
+import { addDoc, deleteDoc, collection, doc, getDoc, getDocs, updateDoc, onSnapshot, query, where, serverTimestamp, setDoc } from 'firebase/firestore'
+import { getCurrentUser, type UserProps } from "./authServices";
+import { deletePrototype } from "./prototypeServices";
 
-// ----- PROJECT RELATED FUNCTIONS -----
-
-// ----- Function to create a new project, that stands for a new machine ----- 
-export const attachProjectToUser = async ( projectId: string, projectName: string, projectsDescription: string, userId: string ) => {
-    try 
-    {
-        const docRef = doc(db, "users", userId, "projects", projectId);
-
-        await setDoc(docRef, {
-            name: projectName,
-            description: projectsDescription,
-            addedToUser: new Date(),
-        });
-
-        return { success: true };
-    }
-    catch (err)
-    {
-        console.error(err);
-        return { success: false };
-    }
+// ===== INTERFACE to define type =====
+export interface ProjectProps {
+    id: string;
+    name: string;
+    description: string;
+    ownerId: string;
+    createdAt: string;
 }
 
+// ===== FUNCTIONS =====
+
+// ----- This function creates a new project, that stands for a new machine ----- 
 export const createProject = async ( projectName: string, projectDescription: string, userId: string ) => {
     try
     {
-        const projectsRef = collection(db, "projects");
-        const docRef = await addDoc(projectsRef, { 
+        const collectionRef = collection(db, "projects");
+        const docRef = await addDoc(collectionRef, { 
             name: projectName, 
             description: projectDescription,
-            owner: userId,
+            ownerId: userId,
+            createdAt: serverTimestamp()
+        });
+
+        await setDoc(doc(db, "projectMembers", `${docRef.id}_${userId}`), {
+            projectId: docRef.id,
+            userId,
+            role: "owner",
+            joinedAt: serverTimestamp(),
         });
         
-        // toast.success("✅ Projeto criado com sucesso!");
-
-        attachProjectToUser(docRef.id, projectName, projectDescription, userId);
+        // toast.success("✅ Projeto criado com sucesso!"); 
 
         return docRef.id;
     }
     catch (err)
     {
-        toast.error(`❌ Erro ao criar um novo projeto: ${err}`);
+        toast.error(`❌ Erro ao criar um novo projeto`);
+        console.error(`${err}`);
     }
 }
 
-export const updateProject = async (projectId: string, name: string, description: string) => {
+// ----- This function updates a already created project -----
+export const updateProject = async ( projectId: string, name: string, description: string ) => {
     try 
     {
         const userData = getCurrentUser();
@@ -71,33 +69,52 @@ export const updateProject = async (projectId: string, name: string, description
     }
     catch (err)
     {
-        console.error(err);
+        console.error(`${err}`);
     }
 }
 
-export const getProjects = (callback: (projects: PrototypeProps[]) => void) => {
+export const getProjects = (callback: ( projects: ProjectProps[] ) => void ) => {
     const docRef = collection(db, "projects");
 
     return onSnapshot(docRef, (snapshot) => {
         const projectsData = snapshot.docs.map(doc => ({
             id: doc.id,
             ...doc.data(),
-        }) as PrototypeProps);
+        }) as ProjectProps);
             
         callback(projectsData);
     });
 } 
 
-export const getUserProjects = (userId: string, callback: (projects: PrototypeProps[]) => void) => {
+export const getUserProjects = ( userId: string, callback: ( projects: ProjectProps[] ) => void ) => {
     try 
     {
-        const projectsRef = collection(db, "users", userId, "projects");
+        const collectionRef = collection(db, "projectMembers");
+        const q = query(collectionRef, where("userId", "==", userId));
 
-        const unsubscribe = onSnapshot(projectsRef, (snapshot) => {
-            const projects = snapshot.docs.map((doc) => ({
-                id: doc.id,
-                ...doc.data()
-            }) as PrototypeProps);
+        const unsubscribe = onSnapshot(q, async (snapshot) => {
+            const projectsIds = snapshot.docs.map(doc => doc.data().projectId);
+
+            if(projectsIds.length == 0)
+            {
+                callback([]);
+                return;
+            }
+
+            const projects: ProjectProps[] = [];
+
+            for( const id of projectsIds)
+            {
+                const projectSnap = await getDoc(doc(db, "projects", id));
+
+                if(projectSnap.exists())
+                {
+                    projects.push({
+                        id: projectSnap.id,
+                        ...projectSnap.data()
+                    } as ProjectProps);
+                }
+            }
 
             callback(projects);
         });
@@ -111,7 +128,7 @@ export const getUserProjects = (userId: string, callback: (projects: PrototypePr
     }
 };
 
-export const getProject = async (id: string) => {
+export const getProject = async ( id: string ) => {
     const docRef = doc(db, "projects", id);
     const docSnap = await getDoc(docRef);
 
@@ -126,21 +143,26 @@ export const getProject = async (id: string) => {
     }
 }
 
-export const deleteProjectNPrototypes = async (projectId: string) => {
+export const deleteProjectNPrototypes = async ( projectId: string ) => {
     try
     {
-        // ===== Function to return current user id =====
-
         const user = getCurrentUser();
         const userId = user?.uid;
         if (!userId) return;
 
-        // ===== Once gotten the user id, go to "projects" sun collection and drop the project ref =====
+        const userRole = await getUserRole(projectId);
+        if(userRole !== "owner" || userRole !== "admin")
+        {
+            console.error(`Você não tem permissão!`);
+            return;
+        }
+
+        // ----- Once gotten the user id, go to "projects" collection and drop the project ref -----
 
         const userProjectRef = doc(db, "users", userId, "projects", projectId);
         await deleteDoc(userProjectRef);
 
-        // ===== Delete all prototypes that got project id attache to them ===== 
+        // ----- Delete all prototypes that got project id attach to them ----- 
 
         const prototypesRef = collection(db, "prototypes");
         const q = query(prototypesRef, where("projectId", "==", projectId))
@@ -160,16 +182,163 @@ export const deleteProjectNPrototypes = async (projectId: string) => {
             await Promise.all(deletePromises);
         }
 
-        // ===== Once all prototypes were deleted then delete the projects itself =====
+        // ----- Once all prototypes were deleted then delete the projects itself -----
 
         const projectRef = doc(db, "projects", projectId);
             
         await deleteDoc(projectRef);
 
+        await dropMember(projectId, userId);
+
         toast.info("ℹ️ Projeto e respectivos protótipos excluídos com sucesso!");
     }
     catch (err)
     {
-        toast.error(`❌ Erro ao excluir projeto e seus protótipos: ${err}`);
+        toast.error(`❌ Erro ao excluir projeto e seus protótipos`);
+        console.error(`${err}`);
+    }
+}
+
+// ----- This function links an user with a project -----
+export const linkProjectUser = async ( projectId: string, userId: string, role: string ) => {
+    try 
+    {
+        const docRef = doc(db, "projectMembers", `${projectId}_${userId}`);
+
+        await setDoc(docRef, {
+            projectId,
+            userId,
+            role: role,
+            joinedAt: serverTimestamp(),
+        });
+
+        toast.success(`✅ Usuário adicionado com sucesso ao projeto!`);
+        return { success: true };
+    }
+    catch (err)
+    {
+        console.error(`${err}`);
+        return { success: false };
+    }
+}
+
+// ---- This function returns a list of the members the current project got -----  
+export const getProjectMembers = ( projectId: string, callback: ( users: UserProps[] ) => void ) => {
+    const q = query(collection(db, "projectMembers"), where("projectId", "==", projectId));
+
+    const unsubscribe = onSnapshot(q, (snapshot) => {
+        const membersList = snapshot.docs.map(doc => ({
+            id: doc.id,
+            ...doc.data(),
+        }) as UserProps);;
+
+        callback(membersList);
+    });
+
+    return unsubscribe;
+}
+
+// ----- This function returns a list of users that do not belong to the current project ----- 
+export const getUsersNotInProject = ( projectId: string, callback: ( users: UserProps[] ) => void ) => {
+    const q = query(collection(db, "projectMembers"), where("projectId", "==", projectId));
+
+    const unsubscribe = onSnapshot(q, async (membersSnap) => {
+        const membersIds = membersSnap.docs.map(doc => doc.data().userId);
+
+        const usersSnap = await getDocs(collection(db, "users"));
+
+        const users = usersSnap.docs.map(doc => ({
+            id: doc.id,
+            ...doc.data()
+        })).filter(user => !membersIds.includes(user.id));
+
+        callback(users as UserProps[]);
+    });
+
+    return unsubscribe;
+}
+
+// ----- This function gets and returns the user role -----
+export const getUserRole = async ( projectId: string ) => {
+    try
+    {
+        const userData = getCurrentUser();
+        if(!userData)
+        {
+            console.error(`Erro na tentativa de puxar informações do usuário!`);
+            return null;
+        }
+
+        const userId = userData?.uid;
+
+        const docRef = doc(db, "projectMembers", `${projectId}_${userId}`);
+        const docSnap = await getDoc(docRef);
+
+        if(!docSnap.exists()) return null;
+
+        return docSnap.data().role ?? null;
+    }
+    catch(err)
+    {
+        console.error(`${err}`);
+        return null;
+    }
+}
+
+// ----- This function changes the role of a project member -----
+export const changeMemberRole = async ( projectId: string, userId: string, role: string ) => {
+    try 
+    {
+        const userRole = await getUserRole(projectId);
+        if(userRole !== "owner" || userRole !== "admin")
+        {
+            console.error(`Você não tem permissão!`);
+            return;
+        }
+
+        const q = query(collection(db, "projectMembers"), where("projectId", "==", projectId), where("userId", "==", userId));
+        const querySnapshot = await getDocs(q);
+
+        if(!querySnapshot.empty)
+        {
+            const docSnap = querySnapshot.docs[0];
+            const docRef = docSnap.ref;
+
+            await updateDoc(docRef, {
+                role: role
+            });
+        }
+    }
+    catch(err)
+    {
+        console.error(`${err}`);
+    }
+}
+
+// ----- This function drops out a member from a project -----
+export const dropMember = async ( projectId: string, userId: string ) => {
+    try 
+    {
+        const userRole = await getUserRole(projectId);
+        if(userRole !== "owner" || userRole !== "admin")
+        {
+            console.error(`Você não tem permissão!`);
+            return;
+        }
+
+        const q = query(collection(db, "projectMembers"), where("projectId", "==", projectId), where("userId", "==", userId));
+        const querySnapshot = await getDocs(q);
+
+        if(!querySnapshot.empty)
+        {
+            const docSnap = querySnapshot.docs[0];
+            const docRef = docSnap.ref;
+
+            await deleteDoc(docRef);
+        }
+    }
+    catch(err)
+    {
+        console.error(`${err}`);
     }
 }
